+++ src/frequenz/sdk/microgrid/component/_component.py
@@ -2,16 +2,19 @@
 # Copyright © 2022 Frequenz Energy-as-a-Service GmbH
 
 """Defines the components that can be used in a microgrid."""
+from __future__ import annotations
 
 from dataclasses import dataclass
 from enum import Enum
+from typing import TYPE_CHECKING
 
 import frequenz.api.common.components_pb2 as components_pb
 import frequenz.api.microgrid.grid_pb2 as grid_pb
 import frequenz.api.microgrid.inverter_pb2 as inverter_pb
 
-from ...timeseries import Current
-from ..fuse import Fuse
+if TYPE_CHECKING:
+    # Break circular import
+    from ...timeseries import Fuse
 
 
 class ComponentType(Enum):
@@ -140,6 +143,8 @@ def _component_metadata_from_protobuf(
     component_category: components_pb.ComponentCategory.ValueType,
     component_metadata: grid_pb.Metadata,
 ) -> GridMetadata | None:
+    from ...timeseries import Current, Fuse  # pylint: disable=import-outside-toplevel
+
     if component_category == components_pb.ComponentCategory.COMPONENT_CATEGORY_GRID:
         max_current = Current.from_amperes(component_metadata.rated_fuse_current)
         fuse = Fuse(max_current)